capture progress on sitemap
