+++ typescript/apps/sage-backend/package.json
@@ -16,12 +16,14 @@
     "@notionhq/client": "^4.0.1",
     "@pinecone-database/pinecone": "^6.1.2",
     "@slack/web-api": "^7.9.3",
+    "async-sema": "^3.1.1",
     "cheerio": "^1.1.2",
     "lodash": "4.17.21",
     "next": "15.4",
     "openai": "4.87.3",
     "react": "19.1.0",
     "react-dom": "19.1.0",
+    "url-join": "^5.0.0",
     "vercel": "^44.5.5",
     "vitest": "^3.2.4",
     "yaml": "^2.8.0",
